+++ src/main/java/webserver/request/RequestHandler.java
@@ -1,6 +1,5 @@
-package webserver;
+package webserver.request;
 
-import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -11,6 +10,7 @@ import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webserver.loader.ResourceLoader;
+import webserver.response.ResponseHandler;
 
 public class RequestHandler implements Runnable {
     public static final String HTTP_METHOD = "Method";
@@ -21,11 +21,13 @@ public class RequestHandler implements Runnable {
     private final Socket connection;
     private final ResourceLoader resourceLoader;
     private final RequestParser requestParser;
+    private final ResponseHandler responseHandler;
 
-    public RequestHandler(Socket connectionSocket, ResourceLoader resourceLoader, RequestParser requestParser) {
+    public RequestHandler(Socket connectionSocket, ResourceLoader resourceLoader, RequestParser requestParser, ResponseHandler responseHandler) {
         this.connection = connectionSocket;
         this.resourceLoader = resourceLoader;
         this.requestParser = requestParser;
+        this.responseHandler = responseHandler;
     }
 
     public void run() {
@@ -36,15 +38,13 @@ public class RequestHandler implements Runnable {
             Map<String, List<String>> requestMap = requestParser.parseRequest(in);
             byte[] body = generateBody(requestMap);
 
-            String type = ContentTypeResolver.getContentType(requestMap.get(REQUEST_URL).getFirst());
-            DataOutputStream dos = new DataOutputStream(out);
-            response200Header(dos, body.length, type);
-            responseBody(dos, body);
+            responseHandler.createResponse(requestMap, out, body);
         } catch (IOException e) {
             logger.error(e.getMessage());
         }
 
     }
+
 
     private byte[] generateBody(Map<String, List<String>> requestMap) {
         byte[] body = new byte[0];
@@ -54,24 +54,4 @@ public class RequestHandler implements Runnable {
         return body;
     }
 
-
-    private void response200Header(DataOutputStream dos, int lengthOfBodyContent, String type) {
-        try {
-            dos.writeBytes("HTTP/1.1 200 OK \r\n");
-            dos.writeBytes("Content-Type: " + type + "\r\n");
-            dos.writeBytes("Content-Length: " + lengthOfBodyContent + "\r\n");
-            dos.writeBytes("\r\n");
-        } catch (IOException e) {
-            logger.error(e.getMessage());
-        }
-    }
-
-    private void responseBody(DataOutputStream dos, byte[] body) {
-        try {
-            dos.write(body, 0, body.length);
-            dos.flush();
-        } catch (IOException e) {
-            logger.error(e.getMessage());
-        }
-    }
 }
