+++ src/main/java/webserver/response/ResponseHandler.java
@@ -1,0 +1,44 @@
+package webserver.response;
+
+import static webserver.request.RequestHandler.REQUEST_URL;
+
+import java.io.DataOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.util.List;
+import java.util.Map;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import webserver.ContentTypeResolver;
+
+public class ResponseHandler {
+    private static final Logger logger = LoggerFactory.getLogger(ResponseHandler.class);
+
+    public void createResponse(Map<String, List<String>> requestMap, OutputStream out, byte[] body) {
+        DataOutputStream dos = new DataOutputStream(out);
+        String type = ContentTypeResolver.getContentType(requestMap.get(REQUEST_URL).getFirst());
+
+        response200Header(dos, body.length, type);
+        responseBody(dos, body);
+    }
+
+    private void response200Header(DataOutputStream dos, int lengthOfBodyContent, String type) {
+        try {
+            dos.writeBytes("HTTP/1.1 200 OK \r\n");
+            dos.writeBytes("Content-Type: " + type + "\r\n");
+            dos.writeBytes("Content-Length: " + lengthOfBodyContent + "\r\n");
+            dos.writeBytes("\r\n");
+        } catch (IOException e) {
+            logger.error(e.getMessage());
+        }
+    }
+
+    private void responseBody(DataOutputStream dos, byte[] body) {
+        try {
+            dos.write(body, 0, body.length);
+            dos.flush();
+        } catch (IOException e) {
+            logger.error(e.getMessage());
+        }
+    }
+}
